Add seal.py (#60)

+++ src/cmd/seal.sh
@@ -19,20 +19,16 @@ cmd_seal() {
         rm -f "$file.sealed"
     fi
 
-    # get recipients
     get_gpg_key
-    local recipients="--recipient $GPG_KEY"
-    while [[ -n "$1" ]]; do
-        recipients="$recipients --recipient $1"
-        shift
-    done
-
+    local recipients=("$GPG_KEY" "$@")
+    
     # sign and encrypt
     gnupghome_setup
-    gpg --no-tty --auto-key-locate=local,cert,keyserver,pka \
-        --keyserver "$KEYSERVER" $recipients \
-        --sign --encrypt --armor \
-        --output "$file.sealed" "$file"
+    call_gpg seal.py "$file" "$recipients"
+    
+    local err=$?
+    [[ $err == 0 ]] || fail "Error encrypting $file"
+
     gnupghome_reset
 
     [[ -s "$file.sealed" ]] || rm -f "$file.sealed"
